refactor: separate render and presentation logic
- Create custom hook
